HM Services Page (homepage3). Created child component for hmservicescards. Updated profile image,
